refactor: split simulator classes in separate files
