+++ finalynx/simulator/timeline.py
@@ -3,146 +3,9 @@ from datetime import timedelta
 from typing import List
 from typing import Optional
 
-from finalynx import Portfolio
 from finalynx.portfolio.bucket import Bucket
-from finalynx.portfolio.line import Line
-
-
-class Action:
-    def __init__(self, name: Optional[str] = None) -> None:
-        """Abstract class. An action describes a procedure to change something in the portfolio.
-        For instance, when receiving a salary, an action could add some amount to the main account.
-        """
-        self.name = name if name else self.__class__.__name__
-
-    def apply(self, portfolio: Portfolio) -> List["Event"]:
-        """Apply this action's consequence, must be overridden."""
-        raise NotImplementedError("Must be overridden.")
-
-    def __str__(self) -> str:
-        return self.name
-
-
-class SetLineAmount(Action):
-    def __init__(self, target_line: Line, amount: float) -> None:
-        self.target_line = target_line
-        self.amount = amount
-        super().__init__()
-
-    def apply(self, portfolio: Portfolio) -> List["Event"]:
-        self.target_line.amount = self.amount
-        return []
-
-
-class AddLineAmount(Action):
-    def __init__(self, target_line: Line, amount: float) -> None:
-        self.target_line = target_line
-        self.amount = amount
-        super().__init__()
-
-    def apply(self, portfolio: Portfolio) -> List["Event"]:
-        self.target_line.amount += self.amount
-        return []
-
-
-class Recurrence:
-    def __init__(self, until: Optional[date] = None) -> None:
-        self.until = until
-
-    def next(self, current_date: date) -> Optional[date]:
-        next_date = self._next_date(current_date)
-        if self.until and next_date > self.until:
-            return None
-        return next_date
-
-    def _next_date(self, current_date: date) -> date:
-        raise NotImplementedError("Must be overridden by subclass.")
-
-
-class DeltaRecurrence(Recurrence):
-    def __init__(
-        self,
-        days: Optional[int] = None,
-        months: Optional[int] = None,
-        years: Optional[int] = None,
-        until: Optional[date] = None,
-    ) -> None:
-        super().__init__(until)
-        assert days is not None or months is not None or years is not None, "Set at least one time field"
-
-        days = days if days is not None else 0
-        months = months if months is not None else 0
-        years = years if years is not None else 0
-
-        self._delta = timedelta(days, weeks=4 * months + 52 * years)
-
-    def _next_date(self, current_date: date) -> date:
-        return current_date + self._delta
-
-
-class MonthlyRecurrence(Recurrence):
-    def __init__(
-        self,
-        day_of_the_month: int,
-        until: Optional[date] = None,
-    ) -> None:
-        super().__init__(until)
-        self.day_of_the_month = day_of_the_month
-
-    def _next_date(self, current_date: date) -> date:
-        if current_date.month == 12:
-            return date(current_date.year + 1, 1, self.day_of_the_month)
-        return date(current_date.year, current_date.month + 1, self.day_of_the_month)
-
-
-class Event:
-    def __init__(
-        self,
-        action: Action,
-        planned_date: Optional[date] = None,
-        recurrence: Optional[Recurrence] = None,
-        name: Optional[str] = None,
-    ) -> None:
-        """An event associates a date to an action."""
-        self.name = name if name else action.name
-        self.planned_date = planned_date if planned_date is not None else date.today()
-        self.recurrence = recurrence
-        self.action = action
-
-    def apply(self, portfolio: Portfolio) -> List["Event"]:
-        """Apply this event's consequence. This event can generate several new Events."""
-        new_events = self.action.apply(portfolio)
-
-        # Create a copy of this event with the new date if a recurrence is set
-        if self.recurrence:
-            if next_date := self.recurrence.next(self.planned_date):
-                new_events.append(Event(self.action, next_date, self.recurrence, self.name))
-
-        return new_events
-
-    def __str__(self) -> str:
-        return self.name
-
-
-class Salary(Event):
-    def __init__(
-        self,
-        target_line: Line,
-        income: float,
-        expenses: float = 0,
-        day_of_the_month: int = 1,
-        start_date: Optional[date] = None,
-        end_date: Optional[date] = None,
-        name: str = "Salary",
-    ) -> None:
-        super().__init__(
-            AddLineAmount(target_line, income - expenses),
-            start_date
-            if start_date and start_date > date.today()
-            else MonthlyRecurrence(day_of_the_month).next(date.today()),
-            MonthlyRecurrence(day_of_the_month, until=end_date),
-            name,
-        )
+from finalynx.portfolio.folder import Portfolio
+from finalynx.simulator.events import Event
 
 
 class Timeline:
@@ -241,37 +104,3 @@ class Timeline:
 
     def __str__(self) -> str:
         return f"Timeline at {self.current_date}"
-
-
-"""
-if __name__ == "__main__":
-    from rich import print
-
-    line = Line("My Line", amount=1500)
-    portfolio = Portfolio(children=[line])
-
-    print(portfolio.tree())
-
-    t = Timeline(
-        portfolio,
-        [],
-        [
-            # Event(SetLineAmount(line, 1000), date(2023, 8, 27), DeltaRecurrence(months=1, until=date(2025, 1, 1))),
-            # Event(SetLineAmount(line, 1000), date(2023, 8, 27), MonthlyRecurrence(3, until=date(2024, 1, 1))),
-            # Event(AddLineAmount(line, 1000), date(2023, 8, 27), MonthlyRecurrence(3, until=date(2024, 1, 1))),
-            # Event(AddLineAmount(line, 1000), recurrence=MonthlyRecurrence(3)),
-            Salary(line, 2500, 1500, end_date=date(2024, 11, 30), name="Salaire Stellantis"),
-            Event(AddLineAmount(line, 4000), planned_date=date(2024, 4, 10), name="Prime Stellantis"),
-            Event(AddLineAmount(line, 3500), planned_date=date(2025, 4, 10), name="Prime Stellantis"),
-            Salary(line, 3500, 2000, start_date=date(2025, 1, 1)),
-        ],
-        end_date=date(1998 + 65, 4, 5),
-    )
-
-    t.run()
-    print(portfolio.tree())
-
-    print("\n======\nFinished!")
-    print(t)
-    print(line.render())
-"""
